test: add bats tests for python
