+++ implementations/python/tests/bats/examples/setup_suite.bash
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+setup_suite() {
+  load ../load/base.bash
+  load ../load/base_python.bash
+}
+
+teardown_suite() {
+  load ../load/base.bash
+}
