+++ implementations/python/tests/bats/load/base.bash
@@ -1,0 +1,1 @@
+../../../../rust/ockam/ockam_command/tests/bats/load/base.bash
