+++ implementations/python/tests/bats/load/base_python.bash
@@ -1,0 +1,4 @@
+# Examples directory
+if [[ -z $EXAMPLES_DIR ]]; then
+  export EXAMPLES_DIR=examples
+fi
